Big refactor for python harness utilities
Split into multiple files
