+++ test/features/steps/insert_by_period_steps.py
@@ -1,7 +1,6 @@
 from behave import *
 
-from test import dbtvault_generator
-from test import dbtvault_harness_utils
+from test import dbtvault_generator, step_helpers, dbt_runner
 
 
 @step("I insert by period into the {model_name} {vault_structure} "
@@ -25,10 +24,10 @@ def load_table(context, model_name, vault_structure, period, start_date, stop_da
                                            config=config,
                                            **metadata)
 
-    is_full_refresh = dbtvault_harness_utils.is_full_refresh(context)
+    is_full_refresh = step_helpers.is_full_refresh(context)
 
-    logs = dbtvault_harness_utils.run_dbt_models(mode="run", model_names=[model_name],
-                                                 full_refresh=is_full_refresh)
+    logs = dbt_runner.run_dbt_models(mode="run", model_names=[model_name],
+                                     full_refresh=is_full_refresh)
 
     assert "Completed successfully" in logs
 
@@ -52,10 +51,10 @@ def load_table(context, model_name, vault_structure, period):
                                            config=config,
                                            **metadata)
 
-    is_full_refresh = dbtvault_harness_utils.is_full_refresh(context)
+    is_full_refresh = step_helpers.is_full_refresh(context)
 
-    logs = dbtvault_harness_utils.run_dbt_models(mode="run", model_names=[model_name],
-                                                 full_refresh=is_full_refresh)
+    logs = dbt_runner.run_dbt_models(mode="run", model_names=[model_name],
+                                     full_refresh=is_full_refresh)
 
     assert "Completed successfully" in logs
 
@@ -79,9 +78,9 @@ def load_table(context, start_date, period, model_name, vault_structure):
                                            config=config,
                                            **metadata)
 
-    is_full_refresh = dbtvault_harness_utils.is_full_refresh(context)
+    is_full_refresh = step_helpers.is_full_refresh(context)
 
-    logs = dbtvault_harness_utils.run_dbt_models(mode="run", model_names=[model_name],
-                                                 full_refresh=is_full_refresh)
+    logs = dbt_runner.run_dbt_models(mode="run", model_names=[model_name],
+                                     full_refresh=is_full_refresh)
 
     assert "Completed successfully" in logs
