+++ test/test_harness/test_dbtvault_harness_utils/test_is_pipeline.py
@@ -2,7 +2,6 @@ import os
 from unittest.mock import patch
 
 from env import env_utils
-from test import dbtvault_harness_utils
 
 
 @patch.dict(os.environ, {'PIPELINE_JOB': 'test_job', 'PIPELINE_BRANCH': 'test'}, clear=True)
