+++ test/test_macro/internal/test_alias_all.py
@@ -1,6 +1,6 @@
 import pytest
 
-from test import dbtvault_harness_utils
+from test import dbt_runner, macro_test_helpers
 
 macro_name = "alias_all"
 
@@ -14,13 +14,13 @@ def test_alias_all_correctly_generates_sql_for_full_alias_list_with_prefix(reque
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -33,13 +33,13 @@ def test_alias_all_correctly_generates_sql_for_partial_alias_list_with_prefix(re
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -51,13 +51,13 @@ def test_alias_all_correctly_generates_sql_for_full_alias_list_without_prefix(re
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -69,11 +69,11 @@ def test_alias_all_correctly_generates_sql_for_partial_alias_list_without_prefix
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
