+++ test/test_macro/staging/test_hash_columns.py
@@ -1,6 +1,6 @@
 import pytest
 
-from test import dbtvault_harness_utils
+from test import dbt_runner, macro_test_helpers
 
 macro_name = "hash_columns"
 
@@ -16,12 +16,12 @@ def test_hash_columns_correctly_generates_hashed_columns_for_single_columns(requ
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -40,12 +40,12 @@ def test_hash_columns_correctly_generates_hashed_columns_for_composite_columns_h
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -60,12 +60,12 @@ def test_hash_columns_correctly_generates_hashed_columns_for_composite_columns_n
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -93,12 +93,12 @@ def test_hash_columns_correctly_generates_hashed_columns_for_multiple_composite_
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -116,12 +116,12 @@ def test_hash_columns_correctly_generates_unsorted_hashed_columns_for_composite_
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -158,12 +158,12 @@ def test_hash_columns_correctly_generates_sql_from_yaml(request, generate_model)
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -198,12 +198,12 @@ def test_hash_columns_correctly_generates_sql_with_constants_from_yaml(request, 
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -233,12 +233,12 @@ def test_hash_columns_raises_warning_if_mapping_without_hashdiff(request, genera
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
     warning_message = "You provided a list of columns under a 'columns' key, " \
                       "but did not provide the 'is_hashdiff' flag. Use list syntax for PKs."
