+++ test/test_macro/staging/test_rank_columns.py
@@ -1,6 +1,6 @@
 import pytest
 
-from test import dbtvault_harness_utils
+from test import dbt_runner, macro_test_helpers
 
 macro_name = "rank_columns"
 
@@ -18,12 +18,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_single_columns(requ
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -40,12 +40,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_single_columns_asc(
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -62,12 +62,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_single_columns_desc
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -85,12 +85,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_single_columns_mult
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -107,12 +107,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_single_columns_mult
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -130,12 +130,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_single_columns_mult
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -155,12 +155,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_single_columns_mult
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -178,12 +178,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_single_columns_dens
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -204,12 +204,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_multiple_columns(re
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -230,12 +230,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_multiple_columns_as
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -256,12 +256,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_multiple_columns_as
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -283,12 +283,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_multiple_columns_mu
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -310,12 +310,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_multiple_columns_mu
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -338,12 +338,12 @@ def test_rank_columns_correctly_generates_ranked_columns_for_multiple_columns_de
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -365,10 +365,10 @@ def test_rank_columns_correctly_generates_ranked_columns_for_multiple_columns_so
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
