+++ test/test_macro/staging/test_stage.py
@@ -1,6 +1,6 @@
 import pytest
 
-from test import dbtvault_harness_utils
+from test import dbt_runner, macro_test_helpers
 
 macro_name = "stage"
 
@@ -28,12 +28,12 @@ def test_stage_correctly_generates_sql_from_yaml(request, generate_model):
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -63,12 +63,12 @@ def test_stage_correctly_generates_sql_from_yaml_with_escape1(request, generate_
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -98,12 +98,12 @@ def test_stage_correctly_generates_sql_from_yaml_with_escape2(request, generate_
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -130,12 +130,12 @@ def test_stage_correctly_generates_sql_from_yaml_with_escape3(request, generate_
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -162,12 +162,12 @@ def test_stage_correctly_generates_sql_from_yaml_with_concat(request, generate_m
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -196,12 +196,12 @@ def test_stage_correctly_generates_sql_from_yaml_with_source_style(request, gene
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -234,12 +234,12 @@ def test_stage_correctly_generates_sql_from_yaml_with_ranked(request, generate_m
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -267,12 +267,12 @@ def test_stage_correctly_generates_sql_from_yaml_with_composite_pk(request, gene
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -285,12 +285,12 @@ def test_stage_correctly_generates_sql_for_only_source_columns_from_yaml(request
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -302,12 +302,12 @@ def test_stage_correctly_generates_sql_for_only_source_columns_and_missing_flag_
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -331,12 +331,12 @@ def test_stage_correctly_generates_sql_for_only_hashing_from_yaml(request, gener
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -353,12 +353,12 @@ def test_stage_correctly_generates_sql_for_only_derived_from_yaml(request, gener
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -380,12 +380,12 @@ def test_stage_correctly_generates_sql_for_only_ranked_from_yaml(request, genera
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -408,12 +408,12 @@ def test_stage_correctly_generates_sql_for_hashing_and_source_from_yaml(request,
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -431,12 +431,12 @@ def test_stage_correctly_generates_sql_for_hashing_and_source_from_yaml_for_sing
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -456,12 +456,12 @@ def test_stage_correctly_generates_sql_for_hashing_and_source_from_yaml_for_sing
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -486,12 +486,12 @@ def test_stage_correctly_generates_sql_for_hashing_and_source_from_yaml_for_mult
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -519,12 +519,12 @@ def test_stage_correctly_generates_sql_for_hashing_and_derived_from_yaml(request
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -540,12 +540,12 @@ def test_stage_correctly_generates_sql_for_derived_and_source_from_yaml(request,
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -574,12 +574,12 @@ def test_stage_correctly_generates_sql_for_hashing_with_exclude_flag(request, ge
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -602,12 +602,12 @@ def test_stage_correctly_generates_sql_for_hashing_with_exclude_flag_no_columns(
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -633,12 +633,12 @@ def test_stage_correctly_generates_sql_for_only_hashing_with_exclude_flag(reques
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -663,12 +663,12 @@ def test_stage_correctly_generates_sql_for_only_source_and_hashing_with_exclude_
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -695,7 +695,7 @@ def test_stage_raises_error_with_missing_source(request, generate_model):
 
     generate_model(metadata)
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name])
 
     assert 'Staging error: Missing source_model configuration. ' \
            'A source model name must be provided.' in dbt_logs
