+++ test/test_macro/supporting/test_hash.py
@@ -1,6 +1,6 @@
 import pytest
 
-from test import dbtvault_harness_utils
+from test import dbt_runner, macro_test_helpers
 
 macro_name = "hash"
 
@@ -11,12 +11,12 @@ def test_hash_single_column_is_successful_md5(request, generate_model):
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -26,12 +26,12 @@ def test_hash_single_column_is_successful_sha(request, generate_model):
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -41,12 +41,12 @@ def test_hash_single_item_list_column_for_pk_is_successful(request, generate_mod
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -56,12 +56,12 @@ def test_hash_single_item_list_column_for_hashdiff_is_successful(request, genera
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -71,12 +71,12 @@ def test_hash_multi_column_as_pk_is_successful_md5(request, generate_model):
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -87,12 +87,12 @@ def test_hash_multi_column_as_pk_is_successful_sha(request, generate_model):
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -102,12 +102,12 @@ def test_hash_multi_column_as_pk_is_successful_custom_concat(request, generate_m
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -117,12 +117,12 @@ def test_hash_multi_column_as_pk_is_successful_custom_null(request, generate_mod
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -133,12 +133,12 @@ def test_hash_multi_column_as_pk_is_successful_custom_null_concat(request, gener
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -149,12 +149,12 @@ def test_hash_multi_column_as_hashdiff_is_successful_md5(request, generate_model
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -165,12 +165,12 @@ def test_hash_multi_column_as_hashdiff_is_successful_sha(request, generate_model
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -180,12 +180,12 @@ def test_hash_single_column_expression3_is_successful_md5(request, generate_mode
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -195,10 +195,10 @@ def test_hash_multi_column_expression4_is_successful_md5(request, generate_model
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
