+++ test/test_macro/supporting/test_prefix.py
@@ -1,6 +1,6 @@
 import pytest
 
-from test import dbtvault_harness_utils
+from test import dbt_runner, macro_test_helpers
 
 macro_name = "prefix"
 
@@ -11,12 +11,12 @@ def test_prefix_column_in_single_item_list_is_successful(request, generate_model
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -26,12 +26,12 @@ def test_prefix_multiple_columns_is_successful(request, generate_model):
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -42,12 +42,12 @@ def test_prefix_aliased_column_is_successful(request, generate_model):
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -58,12 +58,12 @@ def test_prefix_aliased_column_with_alias_target_as_source_is_successful(request
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -74,12 +74,12 @@ def test_prefix_aliased_column_with_alias_target_as_target_is_successful(request
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     args=var_dict)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         args=var_dict)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -89,7 +89,7 @@ def test_prefix_with_no_columns_raises_error(request, generate_model):
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name], args=var_dict)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name], args=var_dict)
 
     assert "Invalid parameters provided to prefix macro. Expected: " \
            "(columns [list/string], prefix_str [string]) got: ([], c)" in dbt_logs
@@ -101,7 +101,7 @@ def test_prefix_with_empty_column_list_raises_error(request, generate_model):
 
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name], args=var_dict)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name], args=var_dict)
 
     assert "Invalid parameters provided to prefix macro. Expected: " \
            "(columns [list/string], prefix_str [string]) got: ([], c)" in dbt_logs
