+++ test/test_macro/tables/test_hub.py
@@ -1,6 +1,6 @@
 import pytest
 
-from test import dbtvault_harness_utils
+from test import dbt_runner, macro_test_helpers
 
 macro_name = "hub"
 
@@ -9,13 +9,13 @@ macro_name = "hub"
 def test_hub_macro_correctly_generates_sql_for_single_source(request, generate_model):
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     full_refresh=True)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         full_refresh=True)
 
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -23,12 +23,12 @@ def test_hub_macro_correctly_generates_sql_for_single_source(request, generate_m
 def test_hub_macro_correctly_generates_sql_for_single_source_multi_nk(request, generate_model):
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     full_refresh=True)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         full_refresh=True)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -36,15 +36,15 @@ def test_hub_macro_correctly_generates_sql_for_single_source_multi_nk(request, g
 def test_hub_macro_correctly_generates_sql_for_incremental_single_source(request, generate_model):
     generate_model()
 
-    dbt_logs_first_run = dbtvault_harness_utils.run_dbt_models(mode='run',
-                                                               model_names=[request.node.name],
-                                                               full_refresh=True)
-    dbt_logs_inc_run = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs_first_run = dbt_runner.run_dbt_models(mode='run',
+                                                   model_names=[request.node.name],
+                                                   full_refresh=True)
+    dbt_logs_inc_run = dbt_runner.run_dbt_models(model_names=[request.node.name])
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_first_run)
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_inc_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_first_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_inc_run)
     assert actual_sql == expected_sql
 
 
@@ -52,15 +52,15 @@ def test_hub_macro_correctly_generates_sql_for_incremental_single_source(request
 def test_hub_macro_correctly_generates_sql_for_incremental_single_source_multi_nk(request, generate_model):
     generate_model()
 
-    dbt_logs_first_run = dbtvault_harness_utils.run_dbt_models(mode='run',
-                                                               model_names=[request.node.name],
-                                                               full_refresh=True)
-    dbt_logs_inc_run = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs_first_run = dbt_runner.run_dbt_models(mode='run',
+                                                   model_names=[request.node.name],
+                                                   full_refresh=True)
+    dbt_logs_inc_run = dbt_runner.run_dbt_models(model_names=[request.node.name])
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_first_run)
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_inc_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_first_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_inc_run)
     assert actual_sql == expected_sql
 
 
@@ -68,12 +68,12 @@ def test_hub_macro_correctly_generates_sql_for_incremental_single_source_multi_n
 def test_hub_macro_correctly_generates_sql_for_multi_source(request, generate_model):
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     full_refresh=True)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         full_refresh=True)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -81,12 +81,12 @@ def test_hub_macro_correctly_generates_sql_for_multi_source(request, generate_mo
 def test_hub_macro_correctly_generates_sql_for_multi_source_multi_nk(request, generate_model):
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     full_refresh=True)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         full_refresh=True)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -94,15 +94,15 @@ def test_hub_macro_correctly_generates_sql_for_multi_source_multi_nk(request, ge
 def test_hub_macro_correctly_generates_sql_for_incremental_multi_source(request, generate_model):
     generate_model()
 
-    dbt_logs_first_run = dbtvault_harness_utils.run_dbt_models(mode='run',
-                                                               model_names=[request.node.name],
-                                                               full_refresh=True)
-    dbt_logs_inc_run = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs_first_run = dbt_runner.run_dbt_models(mode='run',
+                                                   model_names=[request.node.name],
+                                                   full_refresh=True)
+    dbt_logs_inc_run = dbt_runner.run_dbt_models(model_names=[request.node.name])
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_first_run)
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_inc_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_first_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_inc_run)
     assert actual_sql == expected_sql
 
 
@@ -110,13 +110,13 @@ def test_hub_macro_correctly_generates_sql_for_incremental_multi_source(request,
 def test_hub_macro_correctly_generates_sql_for_incremental_multi_source_multi_nk(request, generate_model):
     generate_model()
 
-    dbt_logs_first_run = dbtvault_harness_utils.run_dbt_models(mode='run',
-                                                               model_names=[request.node.name],
-                                                               full_refresh=True)
-    dbt_logs_inc_run = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs_first_run = dbt_runner.run_dbt_models(mode='run',
+                                                   model_names=[request.node.name],
+                                                   full_refresh=True)
+    dbt_logs_inc_run = dbt_runner.run_dbt_models(model_names=[request.node.name])
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_first_run)
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_inc_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_first_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_inc_run)
     assert actual_sql == expected_sql
