+++ test/test_macro/tables/test_link.py
@@ -1,6 +1,6 @@
 import pytest
 
-from test import dbtvault_harness_utils
+from test import dbt_runner, macro_test_helpers
 
 macro_name = "link"
 
@@ -9,12 +9,12 @@ macro_name = "link"
 def test_link_macro_correctly_generates_sql_for_single_source(request, generate_model):
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     full_refresh=True)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         full_refresh=True)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -22,15 +22,15 @@ def test_link_macro_correctly_generates_sql_for_single_source(request, generate_
 def test_link_macro_correctly_generates_sql_for_incremental_single_source(request, generate_model):
     generate_model()
 
-    dbt_logs_first_run = dbtvault_harness_utils.run_dbt_models(mode='run',
-                                                               model_names=[request.node.name],
-                                                               full_refresh=True)
-    dbt_logs_inc_run = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs_first_run = dbt_runner.run_dbt_models(mode='run',
+                                                   model_names=[request.node.name],
+                                                   full_refresh=True)
+    dbt_logs_inc_run = dbt_runner.run_dbt_models(model_names=[request.node.name])
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_first_run)
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_inc_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_first_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_inc_run)
     assert actual_sql == expected_sql
 
 
@@ -38,12 +38,12 @@ def test_link_macro_correctly_generates_sql_for_incremental_single_source(reques
 def test_link_macro_correctly_generates_sql_for_multi_source(request, generate_model):
     generate_model()
 
-    dbt_logs = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name],
-                                                     full_refresh=True)
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs = dbt_runner.run_dbt_models(model_names=[request.node.name],
+                                         full_refresh=True)
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs)
+    assert macro_test_helpers.is_successful_run(dbt_logs)
     assert actual_sql == expected_sql
 
 
@@ -51,13 +51,13 @@ def test_link_macro_correctly_generates_sql_for_multi_source(request, generate_m
 def test_link_macro_correctly_generates_sql_for_incremental_multi_source(request, generate_model):
     generate_model()
 
-    dbt_logs_first_run = dbtvault_harness_utils.run_dbt_models(mode='run',
-                                                               model_names=[request.node.name],
-                                                               full_refresh=True)
-    dbt_logs_inc_run = dbtvault_harness_utils.run_dbt_models(model_names=[request.node.name])
-    actual_sql = dbtvault_harness_utils.retrieve_compiled_model(request.node.name)
-    expected_sql = dbtvault_harness_utils.retrieve_expected_sql(request)
+    dbt_logs_first_run = dbt_runner.run_dbt_models(mode='run',
+                                                   model_names=[request.node.name],
+                                                   full_refresh=True)
+    dbt_logs_inc_run = dbt_runner.run_dbt_models(model_names=[request.node.name])
+    actual_sql = macro_test_helpers.retrieve_compiled_model(request.node.name)
+    expected_sql = macro_test_helpers.retrieve_expected_sql(request)
 
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_first_run)
-    assert dbtvault_harness_utils.is_successful_run(dbt_logs_inc_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_first_run)
+    assert macro_test_helpers.is_successful_run(dbt_logs_inc_run)
     assert actual_sql == expected_sql
